Integrated vote page, further fleshed out templates for pages.
